Lock display on startup mask when klippy is not ready
The Display will now show only the "Startup Mask" when Klippy is not ready.

The Startup Mask:

* shows the Klippy State
* the Klippy Message
* has a button for "Restart"
* has a button for "FW Restart"

For the implementation the WebsocketInterface was extended that changes
of the klippy state can be subscribed.

Also changed startup routine that a proper error message is now shown
when reading on of the configuration files has failed.

Fixes #16